projectdemo_pytest-interface-automation/requirements.txt
pytest~=6.2.4
pytest-xdist
pytest-ordering
pytest-rerunfailures
pytest-html
allure-pytest
requests~=2.25.1
PyYAML~=5.4.1
jsonpath~=0.82
selenium~=3.141.0
openpyxl~=3.0.9
loguru~=0.6.0